% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheerfulgif.R
\name{gif_server}
\alias{gif_server}
\title{Cheerful GIF Button Server Logic}
\usage{
gif_server(id)
}
\arguments{
\item{id}{A character string used as the Shiny module namespace ID.}
}
\value{
Server-side logic for displaying a random GIF and message.
}
\description{
Cheerful GIF Button Server Logic
}
