% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheerfulgif.R
\name{gif_ui}
\alias{gif_ui}
\title{Cheerful GIF Button UI}
\usage{
gif_ui(
  id,
  color = "orange",
  textcolor = NULL,
  label = "Cheer me up",
  position = "fixed; bottom: 20px; right: 20px;"
)
}
\arguments{
\item{id}{A character string used as the Shiny module namespace ID.}

\item{color}{Background color of the button. Default is "orange".}

\item{textcolor}{Text color of the button. If NULL, automatically computed for best contrast.}

\item{label}{Text to display on the button. Default is "Cheer me up".}

\item{position}{CSS positioning for the button. Default is bottom-right.}
}
\value{
A Shiny UI button that triggers a cheerful GIF popup.
}
\description{
Cheerful GIF Button UI
}
